Update current format to Vnd, Update logo, icon

+++ src/java/Controller/Customer/AddToOrder.java
@@ -171,7 +171,7 @@ public class AddToOrder extends HttpServlet {
             vnp_Params.put("vnp_Version", vnp_Version);
             vnp_Params.put("vnp_Command", vnp_Command);
             vnp_Params.put("vnp_TmnCode", vnp_TmnCode);
-            vnp_Params.put("vnp_Amount", String.valueOf((int) (order.getTotalcost() * 100)));
+            vnp_Params.put("vnp_Amount", String.valueOf((long) (order.getTotalcost() * 100)));
             vnp_Params.put("vnp_CurrCode", "VND");
 
             if (bankCode != null && !bankCode.isEmpty()) {
